requirements.txt
Flask
asyncio
requests
configparser
office365-runtime
office365-sharepoint
requests